This code defines classes for investments, stocks, bonds, and mutual funds, as well as a class for a portfolio that can contain multiple investments. The Investment class has methods for getting the current price and calculating the current value of the investment using the Alpha Vantage API. The Stock, Bond, and MutualFund classes inherit from the Investment class and add their specific properties. The Portfolio class has methods for adding and removing investments, calculating the total value of the portfolio, and displaying the portfolio data.
The run() function creates an instance of the Portfolio class and adds three investments to it: a stock, a bond, and a mutual fund. These investments have various quantities, purchase prices, and additional properties specific to their type. The display method of the Portfolio class is called to show the current values of all the investments in the portfolio, as well as the total value of the portfolio.